Port to DevilDaggersInfo.Core.Spawnset (WIP)

+++ DevilDaggersSurvivalEditor/Utils/SpawnUtils.cs
@@ -4,7 +4,7 @@ namespace DevilDaggersSurvivalEditor.Utils;
 
 public static class SpawnUtils
 {
-	public const double MaxDelay = 10000;
+	public const float MaxDelay = 10000;
 
 	public const string Format = "0.0000";
